refactor(customer-analytics): address support tickets tab review
Move conversations facade contracts to contracts.py, convert the
support-tickets endpoint to an @action on AccountViewSet, trim comments.

Generated-By: PostHog Code
Task-Id: 66064672-5e27-456d-bf11-bef9e27fd836
