+++ products/customer_analytics/frontend/components/Accounts/accountSupportTicketsLogic.ts
@@ -11,14 +11,12 @@ export interface AccountSupportTicketsLogicProps {
     accountId: string
 }
 
-// `tickets` is null when the load failed (view shows a "couldn't load" state) or hasn't run yet
-// (distinguished from the NOT_LOADED sentinel below); an empty array is a linked account with no tickets.
 export interface AccountSupportTicketsResult {
     tickets: SupportTicketApi[] | null
     loadFailed?: boolean
 }
 
-// Identity used as a "not loaded yet" sentinel by the view — every loaded outcome returns a fresh object.
+// The view compares against this by reference to tell "not loaded yet" from a loaded empty result.
 export const NOT_LOADED: AccountSupportTicketsResult = { tickets: null }
 
 // Generated by kea-typegen. Update if you're an agent, ignore if you're human.
